Second Solution
On the second solution we use the slice method, which allows us to produce an array with the amount of elements we need.

We start by defining out empty array and a index variable starting at zero to help us get a good idea how to slice our array.

To use the slice method we need to create a while loop to make sure